requirements.txt
Flask==2.0.1

google-cloud>=0.27.0

google-cloud-core>=0.27.1

google-cloud-datastore>=1.3.0